Update SongController.java
functionality for lostupdate buttons is THERE, but doesn't work because netbeans does not allow you to click one button and then another button while the first button is running. the point of a lost update is to click one button to read a table value, wait for 10 seconds, and in that 10 seconds, you click another button that updates the same table value you just read, and then when button one wakes up, it updates and reads that same table value so button 2's changes are lost. I can't do this in netbeans because netbeans doesn't allow you to click another gui button when one gui button is running.